implement transfer resumption
here are how the two cases for transfer resumption are handled by the protocol:

- on early receiver disconnect: the sender's connection hangs while receiver is disconnected. after receiver reconnects,
possibly with a Range request indicating where the transfer needs to start from, server returns 409 CONFLICT to the
sender, with the response body json containing the receiver's indicated Range. sender initiates a new POST with the
correct Range and the transfer resumes.

- on early sender disconnect: when the sender reconnects, it also uses a Range request to indicate where it intends to
resume. if the Range doesn't match what the server thinks the receiver needs, it returns 409 CONFLICT with the Range
information and the process continues like above. receiver connection simply hangs until the correct data is available
from the sender.

one optimization is that we keep the last un-relayed chunk of data from the sender, even after the sender reconnects, to
be able to potentially resend it to the receiver when it reconnects, saving some bandwidth for the sender. on second
thought though, this might be useless in the case of sender reconnect, since that last un-relayed chunk will probably be
way behind what the sender sent, because of tcp buffering.

Closes #3.