put session table in mnesia on designated nodes
the session table remains on the local node by default (but still using mnesia) if the session_table_nodes config value
isn't present. otherwise, it specifies a list of nodes to connect to, to find the session table. if the current node is
in the list, it tries to join an mnesia cluster with the other nodes and create the session table.