respond to OPTIONS so browser can show upload progress
For browsers, In order to implement an "upload indicator", they need to
add an event listener to the ajax request. This causes the request to no
longer be considered a "simple request" meaning that an appropriate CORS
permissions must apply to the requested resource.

For now I did the dumbest thing, which only works for the one URL we
have that requires it.

Probably at some point we'll want to do something smarter.