send an HTTP 100 at upload start and 102 at the end
The HTTP 100 response allows load balancers forwarding requests to this server to know that the server is functioning
and will hopefully terminate any response timeout.

The HTTP 102 just lets the client know that the server received the last chunk of data, which is potentially useful to
clients since they can then safely disconnect if the user doesn't want to wait for the receiver to receive the last
chunk of data (which is currently very likely with small files which are transferred in single in-flight chunk by the
server).