fix content upload timing out after 5s when data is pending
start_upload() should not be a blocking call, and should have a time out, as its return value indicates whether the
session process was found and is responding. so this commit adds an additional call, flush_upload(), called immediately
after start_upload(), which will block forever waiting for any pending data to make it to the downloader.