wait for downloader to finish after receiving 409 Conflict at EOF
If the server accepts the entire file upload content and we perform a retry, the server will respond with a 409 Conflict
telling us to resume at one past the end of the file. Since ordinary Content-Ranges must be at least one byte long, we
use the Content-Range 'asterisk' mechanism to indicate that the requested offset is beyond of the content's length,
which allows us to wait for the downloader to finish without sending any data.