Add formatting script
I don't care much about the particulars if you want to change anything,
I primarily just want it to be uniform.

Xcode likes to add indention on empty lines. To avoid thrashing with Xcode, you might want to:
  1. Open preferences
  2. Text Editing > Editing
  3. Ensure "Automatically trim trailing whitespace" is enabled (this is enabled by default)
  3. Ensure "Include whitespace-only lines" is enabled (this *off* by default)

If we want to use the default 2-space indent, you'd also want to
  1. Open preferences
  2. Text Editing > Indentation
  3. change tab/indent from 4 to 2 spaces

But for now I've configured swift-format to to 4 space indents.